Moved the higher-level game management from **TTTFragment.java** to **GameFragment.java**, which will now be responsible for swapping out individual game fragments.
# Changes Files

## GameFragment.java
* Now contains the high level messaging that will be communicating with individual game fragments. Currently there is only one game fragment, **TTTFragment.java**.

## MainActivity.java
* Replaced *Panel.ttt* with *Panel.game* and all mentions of *TTTFragment* with *GameFragment*, to convert to the new system.

## Panel.java
* *Panel.ttt* has been replaced by *Panel.game*, which uses **GameFragment.java** instead of **TTTFragment.java**

## TTTFragment.java
* Removed higher-level code that is now in **GameFragment.java**

## AndroidManifest.xml
* included android:configChanges attribute to prevent the activity from being destroyed and recreated on an orientation change.

## strings.xml
* R.string.ttt has been replaced by R.string.game

# New Files

## fragment_game.xml
* The **GameFragment.java**’s new layout serves as a fragment container for the individual game fragments.
