+++ app/src/main/java/com/pajato/android/gamechat/main/Panel.java
@@ -23,7 +23,7 @@ import android.util.Log;
 
 import com.pajato.android.gamechat.R;
 import com.pajato.android.gamechat.fragment.ChatFragment;
-import com.pajato.android.gamechat.fragment.TTTFragment;
+import com.pajato.android.gamechat.fragment.GameFragment;
 import com.pajato.android.gamechat.fragment.MembersFragment;
 import com.pajato.android.gamechat.fragment.RoomsFragment;
 
@@ -32,7 +32,7 @@ import com.pajato.android.gamechat.fragment.RoomsFragment;
  */
 public enum Panel {
     chat(R.string.chat, R.layout.fragment_chat, ChatFragment.class),
-    ttt(R.string.ttt, R.layout.fragment_ttt, TTTFragment.class),
+    game(R.string.game, R.layout.fragment_game, GameFragment.class),
     members(R.string.members, R.layout.fragment_members, MembersFragment.class),
     rooms(R.string.rooms, R.layout.fragment_rooms, RoomsFragment.class);
 
@@ -92,8 +92,8 @@ public enum Panel {
                 case R.string.chat:
                     fragment = ChatFragment.instantiate(context, name);
                     break;
-                case R.string.ttt:
-                    fragment = TTTFragment.instantiate(context, name);
+                case R.string.game:
+                    fragment = GameFragment.instantiate(context, name);
                     break;
                 case R.string.members:
                     fragment = MembersFragment.instantiate(context, name);
